Deleted models and migrations in the restructure process
