--- src/Models/QuizAttempt.php
@@ -1,23 +1,0 @@
-<?php
-
-namespace Harishdurga\LaravelQuiz\Models;
-
-use Harishdurga\LaravelQuiz\Models\Quiz;
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-
-class QuizAttempt extends Model
-{
-    use HasFactory;
-    protected $guarded = ['id'];
-
-    public function quiz()
-    {
-        return $this->belongsTo(Quiz::class, 'quiz_id');
-    }
-
-    public function participant()
-    {
-        return $this->morphTo(__FUNCTION__, 'participant_type', 'participant_id');
-    }
-}
