--- src/Models/QuizQuestion.php
@@ -1,33 +1,0 @@
-<?php
-
-namespace Harishdurga\LaravelQuiz\Models;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-
-class QuizQuestion extends Model
-{
-    use HasFactory;
-
-    protected $guarded = ['id'];
-
-    public function getTable()
-    {
-        return config('laravel-quiz.table_names.quiz_questions', parent::getTable());
-    }
-
-    public function quiz()
-    {
-        return $this->belongsTo(Quiz::class, 'quiz_id');
-    }
-
-    protected static function newFactory()
-    {
-        return \Harishdurga\LaravelQuiz\Database\Factories\QuizQuestionFactory::new();
-    }
-
-    public function options()
-    {
-        return $this->hasMany(QuizQuestionOption::class, 'quiz_question_id');
-    }
-}
